requirements.txt
wowool-sdk
wowool-english
wowool-dutch
wowool-anonymizer
wowool-semantic-themes
wowool-topic-identifier
wowool-analysis-formatter
wowool-entity-graph
wowool-phones
wowool-quotes
wowool-numbers
wowool-snippet
wowool-sentiments

